+++ tests/test.cpp
@@ -6,7 +6,7 @@ int main() {
   param::Basis basis;
   param::Basis new_basis;
   std::vector<param::math::vec3> bp(4);
-  std::vector<float> model_vertices(9);
+  std::vector<double> model_vertices(9);
 
   for (int i = 0; i < 4; ++i) {
     bp[i].x = i;
